% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethicalai-package.R
\docType{package}
\name{ethicalai-package}
\alias{ethicalai}
\alias{ethicalai-package}
\title{ethicalai: Ethical AI}
\description{
An R package for creating a transparent audit trail for AI use.
}
\author{
\strong{Maintainer}: Michael Dougherty \email{Michael.P.Dougherty@usace.army.mil} (\href{https://orcid.org/0000-0002-1465-5927}{ORCID})

}
\keyword{internal}
