% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sbom_env.R
\name{read_dcf_field}
\alias{read_dcf_field}
\title{Read a field from a DCF file (DESCRIPTION)}
\usage{
read_dcf_field(path, field)
}
\description{
Read a field from a DCF file (DESCRIPTION)
}
\keyword{internal}
