% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducibleai-package.R
\docType{package}
\name{reproducibleai-package}
\alias{reproducibleai}
\alias{reproducibleai-package}
\title{reproducibleai: Reproducible AI}
\description{
An R package of tools to help teams adopt a reproducible workflow for AI-assisted projects. This package contains a set of tools to assist data science professionals adopt efficient workflows during the early stages of AI adoption. We hold the opinionated view that the ethical use of AI requires data scientists to make well-informed, conscious decisions throughout the analysis process. These tools are a work in progress and designed to help our team experiment with developing new methodologies for AI-assisted workflows. These tools help to define a repeatable framework of chat instructions between sessions and across teams. Tools are provided to create a transparent audit trail to evaluate the of AI.
}
\author{
\strong{Maintainer}: Michael Dougherty \email{Michael.P.Dougherty@usace.army.mil} (\href{https://orcid.org/0000-0002-1465-5927}{ORCID})

}
\keyword{internal}
