% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sbom_env.R
\name{lockfile_sha256}
\alias{lockfile_sha256}
\title{Compute sha256 hash of a file}
\usage{
lockfile_sha256(path)
}
\description{
Uses \code{openssl::sha256()} (recommended for portability). This implies \code{openssl}
should be available (Suggests is acceptable if only used in CI / tooling).
}
\keyword{internal}
